feat(common-ssa-rules): cli rules
